requirements.txt
numpy
moviepy
requests
Pillow
opencv-python
matplotlib